move some layout logic to rustc_target::abi::layout
