+++ compiler/rustc_target/Cargo.toml
@@ -6,6 +6,8 @@ edition = "2021"
 [dependencies]
 bitflags = "1.2.1"
 tracing = "0.1"
+rand = "0.8.4"
+rand_xoshiro = "0.6.0"
 serde_json = "1.0.59"
 rustc_data_structures = { path = "../rustc_data_structures", optional = true  }
 rustc_feature = { path = "../rustc_feature", optional = true }
