+++ compiler/rustc_traits/Cargo.toml
@@ -12,6 +12,7 @@ rustc_hir = { path = "../rustc_hir" }
 rustc_index = { path = "../rustc_index" }
 rustc_ast = { path = "../rustc_ast" }
 rustc_span = { path = "../rustc_span" }
+rustc_target = { path = "../rustc_target" }
 chalk-ir = "0.87.0"
 chalk-engine = "0.87.0"
 chalk-solve = "0.87.0"
